requirements.txt
pandas_datareader
streamlit
yahoo_fin